requirements.txt
openai>=2.0.0,<3.0.0
pdfminer.six==20221105
python-dotenv==1.0.0
rich==13.7.0
streamlit==1.32.0
streamlit-extras==0.4.0
streamlit-option-menu==0.3.12
httpx<0.28
pandas>=2.2.0
numpy>=1.26.0
python-docx>=1.1.0
xlsxwriter>=3.2.0